core: Refactor
